requirements.txt
cirq-core==0.13.1
sympy>=1.8
numpy>=1.23.5
scipy>=1.10.1
pandas>=1.5.3
scikit-learn>=1.2.2
matplotlib>=3.7.1
seaborn>=0.12.2
jupyter>=1.0.0
